+++ go-assignment/racingcar/cmd/result.txt
@@ -1,0 +1,4 @@
+우승자: wer 
+(1등)wer:-----
+(2등)hunki:----
+(3등)qwe:----
