--- go-assignment/racingcar/pkg/common.go
@@ -1,14 +1,0 @@
-package pkg
-
-type IoType int
-
-const (
-	Cli IoType = iota
-	File
-	Json
-)
-
-type User struct {
-	Name          string
-	NumberOfTurns int
-}
