+++ go-assignment/racingcar/pkg/user/user.go
@@ -1,0 +1,6 @@
+package user
+
+type User struct {
+	Name          string
+	NumberOfTurns int
+}
